+++ player/src/bin/play.rs
@@ -11,6 +11,8 @@ use std::{
 
 fn main() {
     #[cfg(feature = "winit")]
+    use raw_window_handle::{HasRawDisplayHandle, HasRawWindowHandle};
+    #[cfg(feature = "winit")]
     use winit::{event_loop::EventLoop, window::WindowBuilder};
 
     env_logger::init();
@@ -45,8 +47,11 @@ fn main() {
     let mut command_buffer_id_manager = wgc::hub::IdentityManager::default();
 
     #[cfg(feature = "winit")]
-    let surface =
-        global.instance_create_surface(&window, wgc::id::TypedId::zip(0, 1, wgt::Backend::Empty));
+    let surface = global.instance_create_surface(
+        window.raw_display_handle(),
+        window.raw_window_handle(),
+        wgc::id::TypedId::zip(0, 1, wgt::Backend::Empty),
+    );
 
     let device = match actions.pop() {
         Some(trace::Action::Init { desc, backend }) => {