+++ wgpu-hal/src/dx12/adapter.rs
@@ -488,11 +488,7 @@ impl crate::Adapter<super::Api> for super::Adapter {
                 | crate::TextureUses::COPY_SRC
                 | crate::TextureUses::COPY_DST,
             present_modes,
-            composite_alpha_modes: vec![
-                crate::CompositeAlphaMode::Opaque,
-                crate::CompositeAlphaMode::PreMultiplied,
-                crate::CompositeAlphaMode::PostMultiplied,
-            ],
+            composite_alpha_modes: vec![wgt::CompositeAlphaMode::Opaque],
         })
     }
 }